Rename kvs package to cache, KVS interface to Cache
The package has always been used as a cache (the field on Dewy is
literally named `cache`), so the kvs naming was a vestige of an
earlier design. Renaming to `cache` aligns the type and package names
with how callers think about and reference it.

This is a mechanical rename:
- `kvs/` -> `cache/`, `kvs.go` -> `cache.go`, `kvs_test.go` -> `cache_test.go`
- `package kvs` -> `package cache` across all files
- `kvs.KVS` interface -> `cache.Cache`
- Import paths updated in dewy.go, dewy_test.go, registry/platform.go

No behavior changes.
